requirements.txt
websockets
python-dotenv
pynput
pyaudio